test(data-tracks): bound stream reads; split out delegate-event tests
DataTrackStream only ends when the track is unpublished, so every bare
next() and `for await frame in stream` waited forever if a frame was
lost — on an unreliable channel that turns a failed assertion into a
hung job rather than a red test. Adds bounded next(within:)/collect(_:)
helpers and routes every read through them, which also replaces the two
hand-rolled task-group-versus-sleep races in the lifecycle suite.

DataTrackTests had become the themeless remainder; its four delegate
event tests move to DataTrackDelegateTests, leaving it the receive
path. Every file stays well inside the length limits.
